Split slow tests
